rework the hlsl lexer
